Splits up the massive flow-graph.ts file into many smaller files.
New structure is the following:
* handle-node.ts: defines node and edges for Handles
* particle-node.ts: same for Particles
* slot-node.ts: same for Slots
* flow-graph.ts: now contains only the code for constructing the graph
* graph-traversal.ts: defines BackwardsPath and graph traversal code
* analysis.ts: contains the graph validation code

In a follow up I will split up the flow-graph-tests.ts file too
